commented "- 1" on line 90 and 95 in file draw.c
yeah it was sending -1 as coordinate for x and y, somehow, it didn't crash on -5 or -6, but it seems like it didn't like -1 or -90 when calculating the position to check the stone in the -90 case or -1. No, I have no iddea why it didmt shit on trying to access -5/6. yes it's ugly whithout it, but I don't care, it's uglier than a 0. I only commented it because one of you might find a workaround, though I doubt it. Good luck if you wanna try on it, I gotta go.